+++ libraries/Temperature/Temperature.h
@@ -1,13 +1,15 @@
 #ifndef TEMPERATURE
 #define TEMPERATURE
 
+#include "inttypes.h"
+
 class Temperature{
 
-    int pin_assignment;
+    uint8_t pin_assignment;
     float filtered_V;
 
     public:
-        Temperature(int);
+        Temperature(uint8_t);
         float get_temp_V(){return filtered_V;};
         void take_reading();
